chore: send telemetry and error handling
