+++ packages/fx-core/src/plugins/resource/bot/errors.ts
@@ -6,6 +6,12 @@ import { CreateAppError, CreateSecretError } from "../aad/errors";
 import { ErrorNames, AzureConstants } from "./constants";
 import { Messages } from "./resources/messages";
 import { getDefaultString, getLocalizedString } from "../../../common/localizeUtils";
+import { err, PluginContext, SystemError, UserError } from "@microsoft/teamsfx-api";
+import { FxBotPluginResultFactory as ResultFactory, FxResult } from "./result";
+import { Logger } from "./logger";
+import { telemetryHelper } from "./utils/telemetry-helper";
+import { CommonHostingError } from "../../../common/azure-hosting/hostingError";
+import { ProgressBarFactory } from "./progressBars";
 
 export const ErrorType = {
   USER: "User",
@@ -282,5 +288,89 @@ export class RegisterResourceProviderError extends PluginError {
       ],
       innerError
     );
+  }
+}
+
+export function wrapError(
+  e: InnerError,
+  context: PluginContext,
+  sendTelemetry: boolean,
+  name: string
+): FxResult {
+  let errorMsg = isErrorWithMessage(e) ? e.message : "";
+  const innerError = isPluginError(e) ? e.innerError : undefined;
+  if (innerError) {
+    errorMsg += getLocalizedString(
+      "plugins.bot.DetailedError",
+      isErrorWithMessage(innerError) ? innerError.message : ""
+    );
+    if (isHttpError(innerError)) {
+      if (innerError.response?.data?.errorMessage) {
+        errorMsg += getLocalizedString(
+          "plugins.bot.DetailedErrorReason",
+          innerError.response?.data?.errorMessage
+        );
+      } else if (innerError.response?.data?.error?.message) {
+        // For errors return from Graph API
+        errorMsg += getLocalizedString(
+          "plugins.bot.DetailedErrorReason",
+          innerError.response?.data?.error?.message
+        );
+      }
+    }
+  }
+  Logger.error(errorMsg);
+  if (e instanceof UserError || e instanceof SystemError) {
+    const res = err(e);
+    sendTelemetry && telemetryHelper.sendResultEvent(context, name, res);
+    return res;
+  }
+
+  if (e instanceof PluginError || e instanceof CommonHostingError) {
+    const result =
+      e instanceof PluginError && e.errorType === ErrorType.SYSTEM
+        ? ResultFactory.SystemError(e.name, [e.genMessage(), e.genDisplayMessage()], e.innerError)
+        : ResultFactory.UserError(
+            e.name,
+            [e.genMessage(), e.genDisplayMessage()],
+            e.innerError,
+            e instanceof PluginError ? e.helpLink : ""
+          );
+    sendTelemetry && telemetryHelper.sendResultEvent(context, name, result);
+    return result;
+  } else {
+    // Unrecognized Exception.
+    const UnhandledErrorCode = "UnhandledError";
+    sendTelemetry &&
+      telemetryHelper.sendResultEvent(
+        context,
+        name,
+        ResultFactory.SystemError(
+          UnhandledErrorCode,
+          [
+            getDefaultString("plugins.bot.UnhandledError", errorMsg),
+            getLocalizedString("plugins.bot.UnhandledError", errorMsg),
+          ],
+          isPluginError(e) ? e.innerError : undefined
+        )
+      );
+    return ResultFactory.SystemError(UnhandledErrorCode, [errorMsg, errorMsg], innerError);
+  }
+}
+
+export async function runWithExceptionCatching<T>(
+  context: PluginContext,
+  fn: () => Promise<FxResult>,
+  sendTelemetry: boolean,
+  name: string
+): Promise<FxResult> {
+  try {
+    sendTelemetry && telemetryHelper.sendStartEvent(context, name);
+    const res: FxResult = await fn();
+    sendTelemetry && telemetryHelper.sendResultEvent(context, name, res);
+    return res;
+  } catch (e) {
+    await ProgressBarFactory.closeProgressBar(false); // Close all progress bars.
+    return wrapError(e, context, sendTelemetry, name);
   }
 }
