+++ packages/fx-core/src/plugins/resource/bot/index.ts
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 import {
   AzureSolutionSettings,
-  err,
   Func,
   FxError,
   ok,
@@ -12,24 +11,14 @@ import {
   QTreeNode,
   Result,
   Stage,
-  SystemError,
-  UserError,
 } from "@microsoft/teamsfx-api";
 
-import { FxBotPluginResultFactory as ResultFactory, FxResult } from "./result";
+import { FxResult } from "./result";
 import { TeamsBotImpl } from "./plugin";
 import { ProgressBarFactory } from "./progressBars";
 import { CustomizedTasks, LifecycleFuncNames, ProgressBarConstants } from "./constants";
-import {
-  ErrorType,
-  InnerError,
-  isErrorWithMessage,
-  isHttpError,
-  isPluginError,
-  PluginError,
-} from "./errors";
+import { runWithExceptionCatching } from "./errors";
 import { Logger } from "./logger";
-import { telemetryHelper } from "./utils/telemetry-helper";
 import { BotOptionItem, MessageExtensionItem } from "../../solution";
 import { Service } from "typedi";
 import { ResourcePlugins } from "../../solution/fx-solution/ResourcePluginContainer";
@@ -45,8 +34,6 @@ import {
   createHostTypeTriggerQuestionForVS,
   showNotificationTriggerCondition,
 } from "./question";
-import { getDefaultString, getLocalizedString } from "../../../common/localizeUtils";
-import { CommonHostingError } from "../../../common/azure-hosting/hostingError";
 
 @Service(ResourcePlugins.BotPlugin)
 export class TeamsBot implements Plugin {
@@ -77,7 +64,7 @@ export class TeamsBot implements Plugin {
   public async scaffold(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    const result = await TeamsBot.runWithExceptionCatching(
+    const result = await runWithExceptionCatching(
       context,
       () => this.getImpl(context, true).scaffold(context),
       true,
@@ -92,7 +79,7 @@ export class TeamsBot implements Plugin {
   public async preProvision(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    return await TeamsBot.runWithExceptionCatching(
+    return await runWithExceptionCatching(
       context,
       () => this.getImpl(context).preProvision(context),
       true,
@@ -103,7 +90,7 @@ export class TeamsBot implements Plugin {
   public async provision(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    const result = await TeamsBot.runWithExceptionCatching(
+    const result = await runWithExceptionCatching(
       context,
       () => this.getImpl(context).provision(context),
       true,
@@ -118,7 +105,7 @@ export class TeamsBot implements Plugin {
   public async postProvision(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    return await TeamsBot.runWithExceptionCatching(
+    return await runWithExceptionCatching(
       context,
       () => this.getImpl(context).postProvision(context),
       true,
@@ -129,7 +116,7 @@ export class TeamsBot implements Plugin {
   public async updateArmTemplates(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    return await TeamsBot.runWithExceptionCatching(
+    return await runWithExceptionCatching(
       context,
       () => this.getImpl(context).updateArmTemplates(context),
       true,
@@ -140,7 +127,7 @@ export class TeamsBot implements Plugin {
   public async generateArmTemplates(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    return await TeamsBot.runWithExceptionCatching(
+    return await runWithExceptionCatching(
       context,
       () => this.getImpl(context).generateArmTemplates(context),
       true,
@@ -151,7 +138,7 @@ export class TeamsBot implements Plugin {
   public async preDeploy(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    return await TeamsBot.runWithExceptionCatching(
+    return await runWithExceptionCatching(
       context,
       () => this.getImpl(context).preDeploy(context),
       true,
@@ -162,7 +149,7 @@ export class TeamsBot implements Plugin {
   public async deploy(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    const result = await TeamsBot.runWithExceptionCatching(
+    const result = await runWithExceptionCatching(
       context,
       () => this.getImpl(context).deploy(context),
       true,
@@ -177,7 +164,7 @@ export class TeamsBot implements Plugin {
   public async localDebug(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    const result = await TeamsBot.runWithExceptionCatching(
+    const result = await runWithExceptionCatching(
       context,
       () => this.getImpl(context).localDebug(context),
       false,
@@ -195,7 +182,7 @@ export class TeamsBot implements Plugin {
   public async postLocalDebug(context: PluginContext): Promise<FxResult> {
     Logger.setLogger(context.logProvider);
 
-    return await TeamsBot.runWithExceptionCatching(
+    return await runWithExceptionCatching(
       context,
       () => this.getImpl(context).postLocalDebug(context),
       false,
@@ -210,7 +197,7 @@ export class TeamsBot implements Plugin {
     Logger.setLogger(context.logProvider);
 
     if (stage === Stage.create) {
-      return await TeamsBot.runWithExceptionCatching(
+      return await runWithExceptionCatching(
         context,
         async () => {
           if (isVSProject(context.projectSettings) || context.answers?.platform === Platform.VS) {
@@ -242,7 +229,7 @@ export class TeamsBot implements Plugin {
   ): Promise<Result<QTreeNode | undefined, FxError>> {
     Logger.setLogger(context.logProvider);
 
-    return await TeamsBot.runWithExceptionCatching(
+    return await runWithExceptionCatching(
       context,
       async () => {
         if (
@@ -263,90 +250,6 @@ export class TeamsBot implements Plugin {
       true,
       LifecycleFuncNames.GET_QUETSIONS_FOR_USER_TASK
     );
-  }
-
-  private static wrapError(
-    e: InnerError,
-    context: PluginContext,
-    sendTelemetry: boolean,
-    name: string
-  ): FxResult {
-    let errorMsg = isErrorWithMessage(e) ? e.message : "";
-    const innerError = isPluginError(e) ? e.innerError : undefined;
-    if (innerError) {
-      errorMsg += getLocalizedString(
-        "plugins.bot.DetailedError",
-        isErrorWithMessage(innerError) ? innerError.message : ""
-      );
-      if (isHttpError(innerError)) {
-        if (innerError.response?.data?.errorMessage) {
-          errorMsg += getLocalizedString(
-            "plugins.bot.DetailedErrorReason",
-            innerError.response?.data?.errorMessage
-          );
-        } else if (innerError.response?.data?.error?.message) {
-          // For errors return from Graph API
-          errorMsg += getLocalizedString(
-            "plugins.bot.DetailedErrorReason",
-            innerError.response?.data?.error?.message
-          );
-        }
-      }
-    }
-    Logger.error(errorMsg);
-    if (e instanceof UserError || e instanceof SystemError) {
-      const res = err(e);
-      sendTelemetry && telemetryHelper.sendResultEvent(context, name, res);
-      return res;
-    }
-
-    if (e instanceof PluginError || e instanceof CommonHostingError) {
-      const result =
-        e instanceof PluginError && e.errorType === ErrorType.SYSTEM
-          ? ResultFactory.SystemError(e.name, [e.genMessage(), e.genDisplayMessage()], e.innerError)
-          : ResultFactory.UserError(
-              e.name,
-              [e.genMessage(), e.genDisplayMessage()],
-              e.innerError,
-              e instanceof PluginError ? e.helpLink : ""
-            );
-      sendTelemetry && telemetryHelper.sendResultEvent(context, name, result);
-      return result;
-    } else {
-      // Unrecognized Exception.
-      const UnhandledErrorCode = "UnhandledError";
-      sendTelemetry &&
-        telemetryHelper.sendResultEvent(
-          context,
-          name,
-          ResultFactory.SystemError(
-            UnhandledErrorCode,
-            [
-              getDefaultString("plugins.bot.UnhandledError", errorMsg),
-              getLocalizedString("plugins.bot.UnhandledError", errorMsg),
-            ],
-            isPluginError(e) ? e.innerError : undefined
-          )
-        );
-      return ResultFactory.SystemError(UnhandledErrorCode, [errorMsg, errorMsg], innerError);
-    }
-  }
-
-  private static async runWithExceptionCatching(
-    context: PluginContext,
-    fn: () => Promise<FxResult>,
-    sendTelemetry: boolean,
-    name: string
-  ): Promise<FxResult> {
-    try {
-      sendTelemetry && telemetryHelper.sendStartEvent(context, name);
-      const res: FxResult = await fn();
-      sendTelemetry && telemetryHelper.sendResultEvent(context, name, res);
-      return res;
-    } catch (e) {
-      await ProgressBarFactory.closeProgressBar(false); // Close all progress bars.
-      return TeamsBot.wrapError(e, context, sendTelemetry, name);
-    }
   }
 
   private static isFunctionsHostedBot(context: PluginContext, isScaffold: boolean): boolean {
