+++ packages/fx-core/src/plugins/resource/bot/v2/index.ts
@@ -8,6 +8,7 @@ import {
   FxError,
   Inputs,
   Json,
+  PluginContext,
   ProjectSettings,
   QTreeNode,
   Result,
@@ -33,16 +34,19 @@ import {
 import {
   configureLocalResourceAdapter,
   configureResourceAdapter,
+  convert2PluginContext,
   executeUserTaskAdapter,
   getQuestionsForScaffoldingAdapter,
   getQuestionsForUserTaskAdapter,
   provisionLocalResourceAdapter,
   provisionResourceAdapter,
+  setEnvInfoV1ByStateV2,
 } from "../../utils4v2";
 import { Logger } from "../logger";
 import { PluginBot } from "../resources/strings";
 import { TeamsBotV2Impl } from "./plugin";
-import { ProgressBarFactory } from "../progressBars";
+import { runWithExceptionCatching } from "../errors";
+import { LifecycleFuncNames } from "../constants";
 
 @Service(ResourcePluginsV2.BotPlugin)
 export class BotPluginV2 implements ResourcePlugin {
@@ -66,7 +70,12 @@ export class BotPluginV2 implements ResourcePlugin {
 
   async scaffoldSourceCode(ctx: Context, inputs: Inputs): Promise<Result<Void, FxError>> {
     Logger.setLogger(ctx.logProvider);
-    return catchAndThrow(() => this.impl.scaffoldSourceCode(ctx, inputs));
+    return runWithExceptionCatching(
+      getV1Context(ctx, inputs),
+      () => this.impl.scaffoldSourceCode(ctx, inputs),
+      true,
+      LifecycleFuncNames.SCAFFOLD
+    );
   }
 
   async generateResourceTemplate(
@@ -74,7 +83,12 @@ export class BotPluginV2 implements ResourcePlugin {
     inputs: Inputs
   ): Promise<Result<ResourceTemplate, FxError>> {
     Logger.setLogger(ctx.logProvider);
-    return catchAndThrow(() => this.impl.generateResourceTemplate(ctx, inputs));
+    return runWithExceptionCatching(
+      getV1Context(ctx, inputs),
+      () => this.impl.generateResourceTemplate(ctx, inputs),
+      true,
+      LifecycleFuncNames.GENERATE_ARM_TEMPLATES
+    );
   }
 
   async updateResourceTemplate(
@@ -82,7 +96,12 @@ export class BotPluginV2 implements ResourcePlugin {
     inputs: Inputs
   ): Promise<Result<v2.ResourceTemplate, FxError>> {
     Logger.setLogger(ctx.logProvider);
-    return catchAndThrow(() => this.impl.updateResourceTemplate(ctx, inputs));
+    return runWithExceptionCatching(
+      getV1Context(ctx, inputs),
+      () => this.impl.updateResourceTemplate(ctx, inputs),
+      true,
+      LifecycleFuncNames.GENERATE_ARM_TEMPLATES
+    );
   }
 
   async provisionResource(
@@ -143,7 +162,12 @@ export class BotPluginV2 implements ResourcePlugin {
     tokenProvider: TokenProvider
   ): Promise<Result<Void, FxError>> {
     Logger.setLogger(ctx.logProvider);
-    return catchAndThrow(() => this.impl.deploy(ctx, inputs, envInfo, tokenProvider));
+    return runWithExceptionCatching(
+      getV1Context(ctx, inputs, envInfo),
+      () => this.impl.deploy(ctx, inputs, envInfo, tokenProvider),
+      true,
+      LifecycleFuncNames.DEPLOY
+    );
   }
 
   async getQuestionsForUserTask(
@@ -183,16 +207,10 @@ export class BotPluginV2 implements ResourcePlugin {
   }
 }
 
-async function catchAndThrow<T>(
-  fn: () => Promise<Result<T, FxError>>
-): Promise<Result<T, FxError>> {
-  try {
-    return await fn();
-  } catch (error: unknown) {
-    await ProgressBarFactory.closeProgressBar(false); // Close all progress bars.
-    if (error instanceof Error) {
-      Logger.error(error.message);
-    }
-    return err(error as FxError);
+function getV1Context(ctx: Context, inputs: Inputs, envInfo?: v2.EnvInfoV2): PluginContext {
+  const context = convert2PluginContext(PluginBot.PLUGIN_NAME, ctx, inputs, true);
+  if (envInfo) {
+    setEnvInfoV1ByStateV2(PluginBot.PLUGIN_NAME, context, envInfo);
   }
+  return context;
 }
